requirement.txt
flask 
twilio 
openai
python-dotenv
google-generativeai 
phidata
edge_tts
groq